Initial Commit
This is a python puzzle where a "rover" is given a position and
direction on a grid and we have to process direction changes and
forward movement commands (while staying on the grid, of course).

For lack of a better convention, I'm going with a Flask like app
structure.

I am taking a functional programming approach to this. One, I like it.
Two, it should scale better and be concurrent, as there are should be no
object locks.